+++ pkg/engine/image_engine_test.go
@@ -46,6 +46,6 @@ func TestImageEngine_Process(t *testing.T) {
 	assert.Nil(t, err)
 	assert.Equal(t, res.StatusCode, 200)
 	assert.Equal(t, res.Headers.Get("content-type"), "image/jpeg")
-	assert.Equal(t, res.Headers.Get("etag"), "4a4e9789cc1e902c")
+	assert.Equal(t, res.Headers.Get("etag"), "W/\"4a4e9789cc1e902c\"")
 	assert.Equal(t, res.Headers.Get("x-amz-meta-public-width"), "300")
 }