+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+* 0.2.0
+    * Feature: Allow register custom url parser 
+    * Bugfix: Change ETag to weak ETag
+    * Bugfix: Fixed race when getting object from cache
 * 0.1.0
     * Feature: New kind of transformation presets-query it combines presets and query in on kind
     * Feature: Added gravity to crop transform